FIX: update ORM and alembic revision scripts
